+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Changes
 
+## v0.3.8
+
+- Added files as input for classic reify function
+
+## v0.3.6
+
+- fix show extension to tall apart show of predicates vs
+
 ## v0.1.0
 
 - create project